[Refactor] better naming function
