+++ src/user/controllers/user.controller.spec.ts
@@ -1,6 +1,7 @@
 import { Test, TestingModule } from '@nestjs/testing';
+
 import { UserController } from './user.controller';
-import { UserService } from '../services/user.service';
+import { UserService } from '../services';
 
 describe('UserController', () => {
   let controller: UserController;
