+++ src/user/dto/update-user.dto.ts
@@ -1,4 +1,5 @@
 import { PartialType } from '@nestjs/mapped-types';
+
 import { CreateUserDto } from './create-user.dto';
 
 export class UpdateUserDto extends PartialType(CreateUserDto) {}
